Implement a falcon compatible dataloader for the graphql gem
The out of the box GraphQL::Dataloader::AsyncDataloader subclasses
GraphQL::Dataloader in a way that there are still fibers spawned
manually. This is fine in a threaded appserver like Puma since there
will just be an Async::Reactor spawned on a per request basis and
torn down once the request is finished. However in an appserver which
is based on async itself, such as falcon, the Async::Reactor is already
spawned at process level. Spawning a fiber manually will not inherit
the reactor and thus spawning a new async task within spawns a new reactor
which in turn replaces the thread's fiber scheduler which in turn terminates
the original reactor which in turn kills any async tasks already running on
the original reactor.

This avoids this issue by just implementing the complete dataloader on primitives
from the async gem. On a threaded appserver the behavior is similar, the first
async task within the dataloder will spawn a reactor on the thread which will
live until all work within it is done and thus termiante as the request finishes.
Avoiding to manually spawn a fiber however then allows on an async based appserver
to just attach all async tasks to the main reactor launched by the appserver.

Note: For the purposes of this discussion the async reactor and fiber scheduler are
largely assumed to be the same thing.